remove dependency on msgpack-rpc
* use pficommon as msgpack RPC server
* fix outputting wrong idl that includes pair<>
* change xxx_serv.hpp spec: take server_argv as constructor's argv
* 'jenerator xxx.hpp', and
* 'mpidl cpp xxx.idl -p -n jubatus -o .' is the command to generate codes

+++ src/server/regression_types.hpp
@@ -83,11 +83,11 @@ public:
   
   MSGPACK_DEFINE(sv, nv);  
 
-  std::vector<std::pair<std::string, std::string> > sv;
-  std::vector<std::pair<std::string, float> > nv;
+  std::vector<std::pair<std::string, std::string > > sv;
+  std::vector<std::pair<std::string, double > > nv;
 };
 
-} // namespace msgpack
+} // namespace jubatus
